+++ CHANGELOG.md
@@ -4,11 +4,11 @@
 
 ### New features
 - [Place Autocomplete] Now `PlaceAutocompleteSuggestion` and `PlaceAutocompleteResult` provide new property `etaMinutes`.
-- [Place Autocomplete] Now `PlaceAutocompleteOptions` accepts an additional parameter `navigationProfile` which that determines how the distance and eta will be calculated.
+- [Place Autocomplete] Now `PlaceAutocompleteOptions` accepts an additional parameter `navigationProfile` that determines how the distance and time of arrival (ETA) will be calculated.
 
 ### Breaking changes
 - [CORE] `SearchNavigationProfile` type has been renamed to `NavigationProfile` and moved to `com.mapbox.search.common` package.
-- [UI] Now `PlaceAutocompleteUiAdapter.search()` function accepts one more argument `MavigationProfile`.
+- [Address Autofill] `AddressAutofillSuggestion.result()` function has been removed. Use `AddressAutofill.select(AddressAutofillSuggestion)` instead. Note that developers must call this function when user selects a search suggestion in the UI.
 
 ### Bug fixes
 - [Place Autocomplete] Fixed a bug with missing `PlaceAutocompleteSuggestion.distanceMeters`.